eth_config: Send ARP requests to the second partition in HV2P
FMan is not able to send broadcasts to multiple queues.
One notable impact of this limitation is in Shared-MAC
scenario where ARP requests destined to the second partition
in HV2P are actually routed to the first partition.

The solution is to use the Software Parser's capabilities, which
means that the parser will analize the ARP packets and will route
the ARP requests to the second partition.
To be noticed that this solution only works for ARP request packets
not for general broadcast frames.

Signed-off-by: Cristian Sovaiala <cristian.sovaiala@freescale.com>